+++ reporter/bootstrapping.go
@@ -12,7 +12,7 @@ import (
 func (r *Reporter) Init() {
 	var (
 		btcLatestBlockHash     *chainhash.Hash
-		btcLatestBlockHeight   int32
+		btcLatestBlockHeight   uint64
 		bbnBaseHeight          uint64
 		bbnLatestBlockHash     *chainhash.Hash
 		bbnLatestBlockHeight   uint64
